Move GetTextureDataAsync from inspector to core
