+++ packages/dev/core/src/Misc/textureTools.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import type { BaseTexture } from "core/Materials/Textures/baseTexture";
 import type { InternalTexture } from "../Materials/Textures/internalTexture";
 import { Texture } from "../Materials/Textures/texture";
 import { RenderTargetTexture } from "../Materials/Textures/renderTargetTexture";
@@ -230,6 +231,179 @@ export function FromHalfFloat(value: number): number {
     return (s ? -1 : 1) * Math.pow(2, e - 15) * (1 + f / Math.pow(2, 10));
 }
 
+const ProcessAsync = async (
+    texture: BaseTexture,
+    width: number,
+    height: number,
+    face: number,
+    channels: { R: boolean; G: boolean; B: boolean; A: boolean },
+    lod: number,
+    resolve: (result: Uint8Array) => void,
+    reject: () => void
+): Promise<void> => {
+    const scene = texture.getScene()!;
+    const engine = scene.getEngine();
+
+    let lodPostProcess: PostProcess;
+
+    if (!texture.isCube) {
+        lodPostProcess = new PostProcess("lod", "lod", ["lod", "gamma"], null, 1.0, null, Texture.NEAREST_NEAREST_MIPNEAREST, engine);
+    } else {
+        const faceDefines = ["#define POSITIVEX", "#define NEGATIVEX", "#define POSITIVEY", "#define NEGATIVEY", "#define POSITIVEZ", "#define NEGATIVEZ"];
+        lodPostProcess = new PostProcess("lodCube", "lodCube", ["lod", "gamma"], null, 1.0, null, Texture.NEAREST_NEAREST_MIPNEAREST, engine, false, faceDefines[face]);
+    }
+
+    if (!lodPostProcess.getEffect().isReady()) {
+        // Try again later
+        lodPostProcess.dispose();
+
+        setTimeout(() => {
+            ProcessAsync(texture, width, height, face, channels, lod, resolve, reject);
+        }, 250);
+
+        return;
+    }
+
+    const rtt = new RenderTargetTexture("temp", { width: width, height: height }, scene, false);
+
+    lodPostProcess.onApply = function (effect) {
+        effect.setTexture("textureSampler", texture);
+        effect.setFloat("lod", lod);
+        effect.setBool("gamma", texture.gammaSpace);
+    };
+
+    const internalTexture = texture.getInternalTexture();
+
+    if (rtt.renderTarget && internalTexture) {
+        const samplingMode = internalTexture.samplingMode;
+        if (lod !== 0) {
+            texture.updateSamplingMode(Texture.NEAREST_NEAREST_MIPNEAREST);
+        } else {
+            texture.updateSamplingMode(Texture.NEAREST_NEAREST);
+        }
+
+        scene.postProcessManager.directRender([lodPostProcess], rtt.renderTarget, true);
+        texture.updateSamplingMode(samplingMode);
+
+        // Read the contents of the framebuffer
+        const numberOfChannelsByLine = width * 4;
+        const halfHeight = height / 2;
+
+        //Reading datas from WebGL
+        const bufferView = await engine.readPixels(0, 0, width, height);
+        const data = new Uint8Array(bufferView.buffer, 0, bufferView.byteLength);
+
+        if (!channels.R || !channels.G || !channels.B || !channels.A) {
+            for (let i = 0; i < width * height * 4; i += 4) {
+                // If alpha is the only channel, just display alpha across all channels
+                if (channels.A && !channels.R && !channels.G && !channels.B) {
+                    data[i] = data[i + 3];
+                    data[i + 1] = data[i + 3];
+                    data[i + 2] = data[i + 3];
+                    data[i + 3] = 255;
+                    continue;
+                }
+                let r = data[i],
+                    g = data[i + 1],
+                    b = data[i + 2],
+                    a = data[i + 3];
+                // If alpha is not visible, make everything 100% alpha
+                if (!channels.A) {
+                    a = 255;
+                }
+                // If only one color channel is selected, map both colors to it. If two are selected, the unused one gets set to 0
+                if (!channels.R) {
+                    if (channels.G && !channels.B) {
+                        r = g;
+                    } else if (channels.B && !channels.G) {
+                        r = b;
+                    } else {
+                        r = 0;
+                    }
+                }
+                if (!channels.G) {
+                    if (channels.R && !channels.B) {
+                        g = r;
+                    } else if (channels.B && !channels.R) {
+                        g = b;
+                    } else {
+                        g = 0;
+                    }
+                }
+                if (!channels.B) {
+                    if (channels.R && !channels.G) {
+                        b = r;
+                    } else if (channels.G && !channels.R) {
+                        b = g;
+                    } else {
+                        b = 0;
+                    }
+                }
+                data[i] = r;
+                data[i + 1] = g;
+                data[i + 2] = b;
+                data[i + 3] = a;
+            }
+        }
+
+        //To flip image on Y axis.
+        if ((texture as Texture).invertY || texture.isCube) {
+            for (let i = 0; i < halfHeight; i++) {
+                for (let j = 0; j < numberOfChannelsByLine; j++) {
+                    const currentCell = j + i * numberOfChannelsByLine;
+                    const targetLine = height - i - 1;
+                    const targetCell = j + targetLine * numberOfChannelsByLine;
+
+                    const temp = data[currentCell];
+                    data[currentCell] = data[targetCell];
+                    data[targetCell] = temp;
+                }
+            }
+        }
+
+        resolve(data);
+
+        // Unbind
+        engine.unBindFramebuffer(rtt.renderTarget);
+    } else {
+        reject();
+    }
+
+    rtt.dispose();
+    lodPostProcess.dispose();
+};
+
+/**
+ * Gets the data of the specified texture by rendering it to an intermediate RGBA texture and retreiving the bytes from it.
+ * This is convienent to get 8-bit RGBA values for a texture in a GPU compressed format.
+ * @param texture the source texture
+ * @param width the width of the result, which does not have to match the source texture width
+ * @param height the height of the result, which does not have to match the source texture height
+ * @param face if the texture has multiple faces, the face index to use for the source
+ * @param channels a filter for which of the RGBA channels to return in the result
+ * @param lod if the texture has multiple LODs, the lod index to use for the source
+ * @returns the 8-bit texture data
+ */
+export async function GetTextureDataAsync(
+    texture: BaseTexture,
+    width: number,
+    height: number,
+    face: number = 0,
+    channels: { R: boolean; G: boolean; B: boolean; A: boolean } = { R: true, G: true, B: true, A: true },
+    lod: number = 0
+): Promise<Uint8Array> {
+    return new Promise((resolve, reject) => {
+        if (!texture.isReady() && texture._texture) {
+            texture._texture.onLoadedObservable.addOnce(() => {
+                ProcessAsync(texture, width, height, face, channels, lod, resolve, reject);
+            });
+            return;
+        }
+
+        ProcessAsync(texture, width, height, face, channels, lod, resolve, reject);
+    });
+}
+
 /**
  * Class used to host texture specific utilities
  */
@@ -268,4 +442,17 @@ export const TextureTools = {
      * @returns converted half float
      */
     FromHalfFloat,
+
+    /**
+     * Gets the data of the specified texture by rendering it to an intermediate RGBA texture and retreiving the bytes from it.
+     * This is convienent to get 8-bit RGBA values for a texture in a GPU compressed format.
+     * @param texture the source texture
+     * @param width the width of the result, which does not have to match the source texture width
+     * @param height the height of the result, which does not have to match the source texture height
+     * @param face if the texture has multiple faces, the face index to use for the source
+     * @param channels a filter for which of the RGBA channels to return in the result
+     * @param lod if the texture has multiple LODs, the lod index to use for the source
+     * @returns the 8-bit texture data
+     */
+    GetTextureDataAsync,
 };
